refactor: separate ui layer
