+++ pyproject.toml
@@ -18,7 +18,7 @@ dependencies = [
     "ttkbootstrap>=1.13.8",
 ]
 
-[project.scripts]
+[project.gui-scripts]
 sotool = "sotool:main"
 
 [build-system]
